feat(insights): experiment with visual new insight menu variants
Add a multivariate experiment (new-insight-menu-experiment) on the saved
insights "New" button: control keeps the existing text menu, "chips" shows a
visual card grid with sketched previews plus sub-insight chips on the Trends
(table, map, number, pie) and Funnel (time to convert, conversion trend)
cards, and "grouped" shows a two-column grid grouped by the question being
asked. Sub-insight entries deep-link to seeded queries via
urls.insightNew({ query }) - no schema changes.

Insight type metadata moves from SavedInsights.tsx to insightTypesMetadata.tsx
(re-exported for existing importers) to avoid an import cycle with the new
menu. "saved insights new insight clicked" gains an optional preset_key
property so preset clicks are distinguishable in experiment analysis.

Generated-By: PostHog Code
Task-Id: f748a9d5-5858-4081-b409-c3f93af203b0
